requirements-dev.txt
pytest==8.4.2
ruff==0.14.14
numpy
torch
streamlit
pyyaml
